add(HeroButtonEditCollection.vue): refresh page on collection data indexed
